requirements.txt
tensorflow==2.8
numpy==1.23
protobuf==3.20
pillow==10.4.0
fastapi==0.115.8
